Changed: #825 Remove all warning when compiling Ryzom on Linux (patch provided by Letto)

+++ code/ryzom/client/src/login_progress_post_thread.cpp
@@ -192,7 +192,7 @@ public:
 					}
 					if (newStep)
 					{
-						if (loginStep.Step == LoginStep_Stop)
+						if (loginStep.Step == (uint)LoginStep_Stop)
 						{
 							break;
 						}
